# src/main/resources/schema/Recipeswagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:10090",
  "basePath": "/",
  "tags": [
    {
      "name": "recipes-rest-controller",
      "description": "Recipes Rest Controller"
    }
  ],
  "paths": {
    "/services/recipe/add": {
      "put": {
        "tags": [
          "recipes-rest-controller"
        ],
        "summary": "Save a new Recipe",
        "operationId": "addRecipeUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "recipe",
            "description": "Recipe object to store value",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Recipe"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "201": {
            "description": "Successfully created",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "Json structure is right but value provided is wrong, Please check whether provided  value of category is already saved"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "The resource you were trying to reach is not found"
          },
          "409": {
            "description": "Duplicate entry"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "deprecated": false
      }
    },
    "/services/recipe/category": {
      "post": {
        "tags": [
          "recipes-rest-controller"
        ],
        "summary": "Save a new category",
        "operationId": "saveDataUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "category",
            "description": "name of a category as a stating value",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "201": {
            "description": "Successfully created",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "The resource you were trying to reach is not found"
          },
          "409": {
            "description": "Duplicate entry"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "deprecated": false
      }
    },
    "/services/recipe/filter/{category}": {
      "get": {
        "tags": [
          "recipes-rest-controller"
        ],
        "summary": "Get all categories of Recipes",
        "operationId": "getDataUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "description": "all/id of a category /name of a category",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved",
            "schema": {
              "$ref": "#/definitions/Categories"
            }
          },
          "204": {
            "description": "No content is present",
            "schema": {
              "$ref": "#/definitions/Categories"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "The resource you were trying to reach is not found"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "deprecated": false
      }
    },
    "/services/recipe/{category}": {
      "get": {
        "tags": [
          "recipes-rest-controller"
        ],
        "summary": "Get all Recipes",
        "operationId": "getRecipesUsingGET_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "description": "name of a category as a stating value",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved",
            "schema": {
              "$ref": "#/definitions/Categories"
            }
          },
          "204": {
            "description": "No content is present",
            "schema": {
              "$ref": "#/definitions/Categories"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "The resource you were trying to reach is not found"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "deprecated": false
      }
    },
    "/services/recipe/{category}/{id}": {
      "get": {
        "tags": [
          "recipes-rest-controller"
        ],
        "summary": "Get a Recipe with a specific category and id",
        "operationId": "getRecipesUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "description": "Name of a category as a stating value",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Id of a recipe",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved",
            "schema": {
              "$ref": "#/definitions/Categories"
            }
          },
          "204": {
            "description": "No content is present",
            "schema": {
              "$ref": "#/definitions/Categories"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "The resource you were trying to reach is not found"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "deprecated": false
      }
    }
  },
  "definitions": {
    "Amount": {
      "type": "object",
      "properties": {
        "qty": {
          "type": "string"
        },
        "unit": {
          "type": "string"
        }
      },
      "title": "Amount"
    },
    "Categories": {
      "type": "object",
      "properties": {
        "cat": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Category"
          }
        },
        "results": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "Categories"
    },
    "Category": {
      "type": "object",
      "properties": {
        "categoryName": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "Category"
    },
    "Directions": {
      "type": "object",
      "properties": {
        "step": {
          "type": "string"
        }
      },
      "title": "Directions"
    },
    "Ingredient": {
      "type": "object",
      "properties": {
        "amt": {
          "$ref": "#/definitions/Amount"
        },
        "item": {
          "type": "string"
        }
      },
      "title": "Ingredient"
    },
    "IngredientDiv": {
      "type": "object",
      "properties": {
        "ing": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ingredient"
          }
        },
        "title": {
          "type": "string"
        }
      },
      "title": "IngredientDiv"
    },
    "Ingredients": {
      "type": "object",
      "properties": {
        "ingdiv": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/IngredientDiv"
          }
        }
      },
      "title": "Ingredients"
    },
    "Recipe": {
      "type": "object",
      "properties": {
        "directions": {
          "$ref": "#/definitions/Directions"
        },
        "head": {
          "$ref": "#/definitions/RecipeHead"
        },
        "id": {
          "type": "string"
        },
        "ingredients": {
          "$ref": "#/definitions/Ingredients"
        }
      },
      "title": "Recipe"
    },
    "RecipeHead": {
      "type": "object",
      "properties": {
        "categories": {
          "$ref": "#/definitions/Categories"
        },
        "title": {
          "type": "string"
        },
        "yield": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "RecipeHead"
    },
    "Recipes": {
      "type": "object",
      "properties": {
        "recipes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Recipe"
          }
        },
        "results": {
          "type": "integer",
          "format": "int64"
        },
        "total": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "Recipes"
    }
  }
}